edit cuisine repo

+++ internal/repository/postgres/cuisine_repo.go
@@ -18,20 +18,21 @@ func NewCuisineRepo(pool *pgxpool.Pool) *CuisineRepo {
 
 func (c *CuisineRepo) Create(ctx context.Context, cuisine *domain.Cuisine) error {
 	query := `
-		INSERT INTO cuisine (name, description)
-		VALUES ($1, $2)
+		INSERT INTO cuisine (name, description, user_id)
+		VALUES ($1, $2, $3)
 		RETURNING id
 		`
 	err := c.pool.QueryRow(ctx, query,
 		cuisine.Name,
 		cuisine.Description,
+		cuisine.UserID,
 	).Scan(&cuisine.ID)
 	return err
 }
 
 func (c *CuisineRepo) List(ctx context.Context, userID int64) ([]*domain.Cuisine, error) {
 	query := `
-			SELECT cuisine.id, cuisine.name, cuisine.description
+			SELECT cuisine.id, cuisine.name, cuisine.description, cuisine.user_id
 			FROM cuisine
 			WHERE user_id = $1
 			ORDER BY cuisine.id
@@ -49,6 +50,7 @@ func (c *CuisineRepo) List(ctx context.Context, userID int64) ([]*domain.Cuisine
 			&cuisine.ID,
 			&cuisine.Name,
 			&cuisine.Description,
+			&cuisine.UserID,
 		)
 		if err != nil {
 			return nil, err
@@ -63,7 +65,7 @@ func (c *CuisineRepo) List(ctx context.Context, userID int64) ([]*domain.Cuisine
 
 func (c *CuisineRepo) GetByID(ctx context.Context, id, userID int64) (*domain.Cuisine, error) {
 	query := `
-			SELECT cuisine.id, cuisine.name, cuisine.description
+			SELECT cuisine.id, cuisine.name, cuisine.description, cuisine.user_id
 			FROM cuisine
 			WHERE user_id = $2 AND cuisine.id = $1
 			`
@@ -72,6 +74,7 @@ func (c *CuisineRepo) GetByID(ctx context.Context, id, userID int64) (*domain.Cu
 		&cuisine.ID,
 		&cuisine.Name,
 		&cuisine.Description,
+		&cuisine.UserID,
 	)
 	if err == pgx.ErrNoRows {
 		return nil, domain.ErrNotFound
@@ -84,7 +87,7 @@ func (c *CuisineRepo) GetByID(ctx context.Context, id, userID int64) (*domain.Cu
 
 func (c *CuisineRepo) GetByName(ctx context.Context, name string, userID int64) (*domain.Cuisine, error) {
 	query := `
-		SELECT cuisine.id, cuisine.name, cuisine.description
+		SELECT cuisine.id, cuisine.name, cuisine.description, user_id
 		FROM cuisine
 		WHERE user_id = $2 AND cuisine.name = $1
 		`
@@ -93,6 +96,7 @@ func (c *CuisineRepo) GetByName(ctx context.Context, name string, userID int64) 
 		&cuisine.ID,
 		&cuisine.Name,
 		&cuisine.Description,
+		&cuisine.UserID,
 	)
 	if err == pgx.ErrNoRows {
 		return nil, domain.ErrNotFound
@@ -108,12 +112,13 @@ func (c *CuisineRepo) Update(ctx context.Context, cuisine *domain.Cuisine) error
 		UPDATE cuisine 
 		SET name = $1,
 			description = $2
-		WHERE id = $3
+		WHERE id = $3 AND user_id = $4
 		`
 	result, err := c.pool.Exec(ctx, query,
 		cuisine.Name,
 		cuisine.Description,
 		cuisine.ID,
+		cuisine.UserID,
 	)
 	if err != nil {
 		return err
@@ -124,9 +129,9 @@ func (c *CuisineRepo) Update(ctx context.Context, cuisine *domain.Cuisine) error
 	return nil
 }
 
-func (c *CuisineRepo) Delete(ctx context.Context, id int64) error {
-	query := `DELETE FROM cuisine WHERE id = $1`
-	result, err := c.pool.Exec(ctx, query, id)
+func (c *CuisineRepo) Delete(ctx context.Context, cuisine domain.Cuisine) error {
+	query := `DELETE FROM cuisine WHERE id = $1 AND user_id = $2`
+	result, err := c.pool.Exec(ctx, query, cuisine.ID, cuisine.UserID)
 	if err != nil {
 		return err
 	}
